+++ lib/src/navigator/navigator_types.dart
@@ -21,14 +21,9 @@
 
 import 'package:flutter/widgets.dart';
 
-import '../module/module_context.dart';
-
 /// Signature of page builder with RouteSettings.
 ///
-typedef NavigatorPageBuilder = Widget Function(
-  RouteSettings settings,
-  ModuleContext context,
-);
+typedef NavigatorPageBuilder = Widget Function(RouteSettings settings);
 
 /// Signature of callbacks with generic parameters with type `T`.
 ///